LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/KeyframeEffect/idlharness-expected.txt


FAIL AnimationEffectReadOnly interface: existence and properties of interface object assert_own_property: self does not have own property "AnimationEffectReadOnly" expected property "AnimationEffectReadOnly" missing
FAIL AnimationEffectReadOnly interface object length assert_own_property: self does not have own property "AnimationEffectReadOnly" expected property "AnimationEffectReadOnly" missing
FAIL AnimationEffectReadOnly interface object name assert_own_property: self does not have own property "AnimationEffectReadOnly" expected property "AnimationEffectReadOnly" missing
FAIL AnimationEffectReadOnly interface: existence and properties of interface prototype object assert_own_property: self does not have own property "AnimationEffectReadOnly" expected property "AnimationEffectReadOnly" missing
FAIL AnimationEffectReadOnly interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "AnimationEffectReadOnly" expected property "AnimationEffectReadOnly" missing
FAIL AnimationEffectReadOnly interface: attribute timing assert_own_property: self does not have own property "AnimationEffectReadOnly" expected property "AnimationEffectReadOnly" missing
FAIL AnimationEffectReadOnly interface: operation getComputedTiming() assert_own_property: self does not have own property "AnimationEffectReadOnly" expected property "AnimationEffectReadOnly" missing
FAIL KeyframeEffectReadOnly interface: existence and properties of interface object assert_own_property: self does not have own property "KeyframeEffectReadOnly" expected property "KeyframeEffectReadOnly" missing
FAIL KeyframeEffectReadOnly interface object length assert_own_property: self does not have own property "KeyframeEffectReadOnly" expected property "KeyframeEffectReadOnly" missing
FAIL KeyframeEffectReadOnly interface object name assert_own_property: self does not have own property "KeyframeEffectReadOnly" expected property "KeyframeEffectReadOnly" missing
FAIL KeyframeEffectReadOnly interface: existence and properties of interface prototype object assert_own_property: self does not have own property "KeyframeEffectReadOnly" expected property "KeyframeEffectReadOnly" missing
FAIL KeyframeEffectReadOnly interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "KeyframeEffectReadOnly" expected property "KeyframeEffectReadOnly" missing
FAIL KeyframeEffectReadOnly interface: attribute target assert_own_property: self does not have own property "KeyframeEffectReadOnly" expected property "KeyframeEffectReadOnly" missing
FAIL KeyframeEffectReadOnly interface: attribute iterationComposite assert_own_property: self does not have own property "KeyframeEffectReadOnly" expected property "KeyframeEffectReadOnly" missing
FAIL KeyframeEffectReadOnly interface: attribute composite assert_own_property: self does not have own property "KeyframeEffectReadOnly" expected property "KeyframeEffectReadOnly" missing
FAIL KeyframeEffectReadOnly interface: operation getKeyframes() assert_own_property: self does not have own property "KeyframeEffectReadOnly" expected property "KeyframeEffectReadOnly" missing
FAIL KeyframeEffectReadOnly must be primary interface of new KeyframeEffectReadOnly(null, null) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: KeyframeEffectReadOnly"
FAIL Stringification of new KeyframeEffectReadOnly(null, null) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: KeyframeEffectReadOnly"
FAIL KeyframeEffectReadOnly interface: new KeyframeEffectReadOnly(null, null) must inherit property "target" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: KeyframeEffectReadOnly"
FAIL KeyframeEffectReadOnly interface: new KeyframeEffectReadOnly(null, null) must inherit property "iterationComposite" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: KeyframeEffectReadOnly"
FAIL KeyframeEffectReadOnly interface: new KeyframeEffectReadOnly(null, null) must inherit property "composite" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: KeyframeEffectReadOnly"
FAIL KeyframeEffectReadOnly interface: new KeyframeEffectReadOnly(null, null) must inherit property "getKeyframes()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: KeyframeEffectReadOnly"
FAIL AnimationEffectReadOnly interface: new KeyframeEffectReadOnly(null, null) must inherit property "timing" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: KeyframeEffectReadOnly"
FAIL AnimationEffectReadOnly interface: new KeyframeEffectReadOnly(null, null) must inherit property "getComputedTiming()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: KeyframeEffectReadOnly"
FAIL KeyframeEffect interface: existence and properties of interface object assert_own_property: should inherit from KeyframeEffectReadOnly, but self has no such property expected property "KeyframeEffectReadOnly" missing
FAIL KeyframeEffect interface object length assert_equals: wrong value for KeyframeEffect.length expected 1 but got 2
PASS KeyframeEffect interface object name 
FAIL KeyframeEffect interface: existence and properties of interface prototype object assert_own_property: should inherit from KeyframeEffectReadOnly, but self has no such property expected property "KeyframeEffectReadOnly" missing
PASS KeyframeEffect interface: existence and properties of interface prototype object's "constructor" property 
FAIL KeyframeEffect interface: attribute target assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
PASS KeyframeEffect interface: attribute iterationComposite 
PASS KeyframeEffect interface: attribute composite 
PASS KeyframeEffect interface: operation setKeyframes(object) 
PASS KeyframeEffect must be primary interface of new KeyframeEffect(null, null) 
PASS Stringification of new KeyframeEffect(null, null) 
PASS KeyframeEffect interface: new KeyframeEffect(null, null) must inherit property "target" with the proper type 
PASS KeyframeEffect interface: new KeyframeEffect(null, null) must inherit property "iterationComposite" with the proper type 
PASS KeyframeEffect interface: new KeyframeEffect(null, null) must inherit property "composite" with the proper type 
PASS KeyframeEffect interface: new KeyframeEffect(null, null) must inherit property "setKeyframes(object)" with the proper type 
PASS KeyframeEffect interface: calling setKeyframes(object) on new KeyframeEffect(null, null) with too few arguments must throw TypeError 
PASS KeyframeEffectReadOnly interface: new KeyframeEffect(null, null) must inherit property "target" with the proper type 
PASS KeyframeEffectReadOnly interface: new KeyframeEffect(null, null) must inherit property "iterationComposite" with the proper type 
PASS KeyframeEffectReadOnly interface: new KeyframeEffect(null, null) must inherit property "composite" with the proper type 
PASS KeyframeEffectReadOnly interface: new KeyframeEffect(null, null) must inherit property "getKeyframes()" with the proper type 
PASS AnimationEffectReadOnly interface: new KeyframeEffect(null, null) must inherit property "timing" with the proper type 
PASS AnimationEffectReadOnly interface: new KeyframeEffect(null, null) must inherit property "getComputedTiming()" with the proper type 
